// Frontend/AngularProject/src/app/clients/student/model/studentAssignmentDto.ts
/**
 * Student api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssignmentStatus } from './assignmentStatus';
import { AttachmentDto } from './attachmentDto';

export interface StudentAssignmentDto { 
    id?: number;
    assignmentId?: number;
    studentId?: number;
    status?: AssignmentStatus;
    body?: string;
    score?: number;
    fineScore?: number;
    groupId?: number;
    submitted?: Date;
    lastStatusChange?: Date;
    completed?: Date;
    attachments?: Array<AttachmentDto>;
}